requirements.txt
openai
langchain
pinecone-client
python-dotenv
tiktoken
docx2txt
pypdf
requests
pandas
aiohttp
PyYAML
numpy